# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pire API",
    "description": "Accounts, lesson progress and the AI tutor for pire.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in with email and password.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Wrong email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign out and end the session.",
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Signed out"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The signed-in user.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Signed out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create an account and sign in.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created and signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/progress": {
      "get": {
        "tags": [
          "progress"
        ],
        "summary": "All of the signed-in user's lesson progress.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Progress for every lesson the user has started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonProgress"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Signed out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/progress/{lesson_id}": {
      "put": {
        "tags": [
          "progress"
        ],
        "summary": "Save progress for one lesson.",
        "operationId": "save",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "description": "Lesson id, for example m1.l1",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveProgressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonProgress"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Signed out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/tutor/status": {
      "get": {
        "tags": [
          "tutor"
        ],
        "summary": "Whether the tutor is configured on this server.",
        "operationId": "status",
        "responses": {
          "200": {
            "description": "Tutor availability",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/tutor/{lesson_id}/messages": {
      "get": {
        "tags": [
          "tutor"
        ],
        "summary": "The learner's tutor chat for one lesson.",
        "operationId": "messages",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "description": "Lesson id, for example m1.l2",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The chat so far",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorMessages"
                }
              }
            }
          },
          "401": {
            "description": "Signed out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Unknown lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tutor"
        ],
        "summary": "Forget the learner's tutor chat for one lesson.",
        "operationId": "reset",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "description": "Lesson id, for example m1.l2",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Chat cleared"
          },
          "401": {
            "description": "Signed out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "LessonProgress": {
        "type": "object",
        "required": [
          "lesson_id",
          "beat_index",
          "state",
          "updated_at"
        ],
        "properties": {
          "beat_index": {
            "type": "integer",
            "format": "int32",
            "description": "The beat the learner should resume at."
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "lesson_id": {
            "type": "string",
            "description": "Stable lesson id from the content package, for example \"m1.l1\"."
          },
          "state": {
            "type": "object",
            "description": "Lesson-specific data such as checkpoint results and hints used."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "SaveProgressRequest": {
        "type": "object",
        "required": [
          "beat_index",
          "completed",
          "state"
        ],
        "properties": {
          "beat_index": {
            "type": "integer",
            "format": "int32"
          },
          "completed": {
            "type": "boolean",
            "description": "Mark the lesson finished. Once set, the first completion time is kept."
          },
          "state": {
            "type": "object"
          }
        }
      },
      "TutorMessages": {
        "type": "object",
        "required": [
          "messages"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "The chat as AI SDK UI messages, ready for useChat."
          }
        }
      },
      "TutorStatus": {
        "type": "object",
        "required": [
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "auth",
      "description": "Accounts and sessions"
    },
    {
      "name": "progress",
      "description": "Lesson progress"
    },
    {
      "name": "tutor",
      "description": "The AI tutor. Its chat stream is POST /api/tutor/{lesson_id}/chat (AI SDK UI message stream)."
    }
  ]
}